+++ src/main/java/com/piotrszustak/quizApp/dto/GameOptionsDto.java
@@ -1,4 +1,4 @@
-package com.piotrszustak.quizApp.dtos;
+package com.piotrszustak.quizApp.dto;
 
 import com.piotrszustak.quizApp.enums.Difficulty;
 import lombok.Getter;
