+++ src/main/java/com/piotrszustak/quizApp/dto/UsersAnswerDto.java
@@ -1,4 +1,4 @@
-package com.piotrszustak.quizApp.dtos;
+package com.piotrszustak.quizApp.dto;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
