+++ src/main/java/com/piotrszustak/quizApp/service/GameService.java
@@ -1,8 +1,8 @@
-package com.piotrszustak.quizApp.services;
+package com.piotrszustak.quizApp.service;
 
-import com.piotrszustak.quizApp.dtos.GameOptionsDto;
-import com.piotrszustak.quizApp.dtos.QuestionsDto;
-import com.piotrszustak.quizApp.dtos.UsersAnswerDto;
+import com.piotrszustak.quizApp.dto.GameOptionsDto;
+import com.piotrszustak.quizApp.dto.QuestionsDto;
+import com.piotrszustak.quizApp.dto.UsersAnswerDto;
 import lombok.Getter;
 import org.springframework.stereotype.Service;
