+++ src/main/java/com/piotrszustak/quizApp/service/QuestionService.java
@@ -1,8 +1,8 @@
-package com.piotrszustak.quizApp.services;
+package com.piotrszustak.quizApp.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.piotrszustak.quizApp.dtos.GameOptionsDto;
-import com.piotrszustak.quizApp.dtos.QuestionsDto;
+import com.piotrszustak.quizApp.dto.GameOptionsDto;
+import com.piotrszustak.quizApp.dto.QuestionsDto;
 import lombok.extern.java.Log;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.HttpClientErrorException;
